# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koumbaya Marketplace API",
        "description": "API REST pour la marketplace Koumbaya avec système de tombolas et paiements mobile money",
        "contact": {
            "name": "Équipe Koumbaya",
            "email": "dev@koumbaya.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement Koumbaya"
        }
    ],
    "paths": {
...
        "/api/payments/initiate": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Initier un paiement pour achat de tickets ou produit",
                "operationId": "09337e33e2aff50f10b00befaf0ae25f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "lottery_ticket",
                                            "product_purchase"
                                        ],
                                        "example": "lottery_ticket"
                                    },
                                    "lottery_id": {
                                        "description": "Required for lottery_ticket type",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_id": {
                                        "description": "Required for product_purchase type",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "description": "Number of tickets (for lottery_ticket)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "phone": {
                                        "description": "Phone for payment",
                                        "type": "string",
                                        "example": "074123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paiement initié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "bill_id": {
                                                    "type": "string"
                                                },
                                                "reference": {
                                                    "type": "string"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/callback": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Callback E-Billing (Webhook)",
                "operationId": "0b732d8a5ed20f67b03a5e8b20267c0d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reference": {
                                        "type": "string"
                                    },
                                    "transactionid": {
                                        "type": "string"
                                    },
                                    "paymentsystem": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Callback traité"
                    }
                }
            }
        },
        "/api/payments/{id}/status": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Vérifier le statut d'un paiement",
                "operationId": "adc18ab90fbf9829221668152a1b1dfc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du paiement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut du paiement"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/success": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Page de succès de paiement",
                "operationId": "ef11379b5af0c382479811f7ee3a2fb2",
                "parameters": [
                    {
                        "name": "reference",
                        "in": "query",
                        "description": "Référence du paiement",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page de succès"
                    }
                }
            }
        },
...
        "/api/payments/ussd-push": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Déclencher un push USSD",
                "operationId": "f27fc5a34e475b3e862c434d2c05cbe1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bill_id",
                                    "payment_system_name",
                                    "payer_msisdn"
                                ],
                                "properties": {
                                    "bill_id": {
                                        "type": "string",
                                        "example": "BILL123456"
                                    },
                                    "payment_system_name": {
                                        "type": "string",
                                        "example": "airtel_money"
                                    },
                                    "payer_msisdn": {
                                        "type": "string",
                                        "example": "074123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Push USSD envoyé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/kyc": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Récupérer les informations KYC",
                "operationId": "946c715356c8c637ddae58608c6e16b1",
                "parameters": [
                    {
                        "name": "operator",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "airtel_money"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "074123456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Données KYC récupérées"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via token Bearer JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}